+++ lib/waterdrop/contractable/rule.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module WaterDrop
+  module Contractable
+    # Representation of a single validation rule
+    Rule = Struct.new(:path, :type, :validator)
+  end
+end