+++ lib/waterdrop/contracts/message.rb
@@ -4,36 +4,47 @@ module WaterDrop
   module Contracts
     # Contract with validation rules for validating that all the message options that
     # we provide to producer ale valid and usable
-    class Message < Base
+    class Message < Contractable::Contract
       # Regex to check that topic has a valid format
       TOPIC_REGEXP = /\A(\w|-|\.)+\z/
 
-      # Checks, that the given value is a string
-      STRING_ASSERTION = ->(value) { value.is_a?(String) }.to_proc
+      private_constant :TOPIC_REGEXP
 
-      private_constant :TOPIC_REGEXP, :STRING_ASSERTION
+      attr_reader :max_payload_size
 
-      option :max_payload_size
-
-      params do
-        required(:topic).filled(:str?, format?: TOPIC_REGEXP)
-        required(:payload).filled(:str?)
-        optional(:key).maybe(:str?, :filled?)
-        optional(:partition).filled(:int?, gteq?: -1)
-        optional(:partition_key).maybe(:str?, :filled?)
-        optional(:timestamp).maybe { time? | int? }
-        optional(:headers).maybe(:hash?)
+      # @param max_payload_size [Integer] max payload size
+      def initialize(max_payload_size:)
+        @max_payload_size = max_payload_size
       end
 
-      rule(:headers) do
-        next unless value.is_a?(Hash)
+      required(:topic) { |topic| topic.is_a?(String) && TOPIC_REGEXP.match?(topic) }
+      required(:payload) { |payload| payload.is_a?(String) }
+      optional(:key) { |key| key.nil? || (key.is_a?(String) && !key.empty?) }
+      optional(:partition) { |partition| partition.is_a?(Integer) && partition >= -1 }
+      optional(:partition_key) { |p_key| p_key.nil? || (p_key.is_a?(String) && !p_key.empty?) }
+      optional(:timestamp) { |ts| ts.nil? || (ts.is_a?(Time) || ts.is_a?(Integer)) }
+      optional(:headers) { |headers| headers.nil? || headers.is_a?(Hash) }
 
-        key.failure(:invalid_key_type) unless value.keys.all?(&STRING_ASSERTION)
-        key.failure(:invalid_value_type) unless value.values.all?(&STRING_ASSERTION)
+      virtual do |config, errors|
+        next true unless errors.empty?
+        next true unless config.key?(:headers)
+        next true if config[:headers].nil?
+
+        errors = []
+
+        config.fetch(:headers).each do |key, value|
+          errors << [%i[headers], :invalid_key_type] unless key.is_a?(String)
+          errors << [%i[headers], :invalid_value_type] unless value.is_a?(String)
+        end
+
+        errors
       end
 
-      rule(:payload) do
-        key.failure(:max_payload_size) if value.bytesize > max_payload_size
+      virtual do |config, errors, validator|
+        next true unless errors.empty?
+        next true if config[:payload].bytesize <= validator.max_payload_size
+
+        [[%i[payload], :max_size]]
       end
     end
   end